# docs/openapi-core-functions.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SightIndex 核心接口：上传、检索、人脸库",
    "version": "2026-09-09-semantic-search-v1",
    "description": "从北京已部署OpenAPI裁剪，配套api-core-functions.zh-CN.md。核对时间2026-09-09 09:52+08:00。保留原请求/成功响应合同；中文摘要、Basic鉴权和部分错误说明为文档补充，不是新增服务能力。新上传数据自动Qwen索引当前关闭；属性生成模型未启用；人员编辑/删除、人脸删除尚无接口。"
  },
  "servers": [
    {
      "url": "http://192.168.1.40:18030",
      "description": "北京内网示例；导入后替换为调用方实际可达地址，不代表公网可达"
    }
  ],
...
  "security": [
    {
      "basicAuth": []
    }
  ],
  "paths": {
...
    "/api/images": {
      "get": {
        "tags": [
          "上传"
        ],
        "summary": "图片列表",
        "operationId": "list_images_api_images_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "has_crops",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Has Crops"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImageRead"
                  },
                  "title": "Response List Images Api Images Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Basic认证失败；当前响应可为纯文本，并带WWW-Authenticate。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "返回裸数组；limit 1～200，支持 offset 和 has_crops。"
      }
    },
...
    "/api/person-crops": {
      "get": {
        "tags": [
          "上传"
        ],
        "summary": "人物裁剪列表",
        "operationId": "list_person_crops_api_person_crops_get",
        "parameters": [
          {
            "name": "image_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Image Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonCropRead"
                  },
                  "title": "Response List Person Crops Api Person Crops Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Basic认证失败；当前响应可为纯文本，并带WWW-Authenticate。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "返回裸数组，可按 image_id 查询。"
      }
    },
...
    "/api/search/semantic/status": {
      "get": {
        "tags": [
          "检索"
        ],
        "summary": "语义检索状态与覆盖",
        "operationId": "semantic_search_status_api_search_semantic_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemanticSearchStatus"
                }
              }
            }
          },
          "401": {
            "description": "Basic认证失败；当前响应可为纯文本，并带WWW-Authenticate。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "configured 是配置条件，indexed_crops 是 SQL 索引标记覆盖，均不是模型及向量库实时健康保证。"
      }
    },
    "/api/search/semantic/person-crops": {
      "post": {
        "tags": [
          "检索"
        ],
        "summary": "Qwen 自然语言检索人物候选",
        "operationId": "semantic_search_person_crops_api_search_semantic_person_crops_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SemanticSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemanticSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Basic认证失败；当前响应可为纯文本，并带WWW-Authenticate。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "业务参数无效或媒体无法处理；具体原因见detail。",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "object",
                          "additionalProperties": true
                        },
                        {
                          "type": "array",
                          "items": {}
                        }
                      ]
                    }
                  },
                  "required": [
                    "detail"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "模型/索引未就绪或不可用、处理队列压力等。视频可能有已提交进度；不能把此状态当作正常空匹配。",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "object",
                          "additionalProperties": true
                        },
                        {
                          "type": "array",
                          "items": {}
                        }
                      ]
                    }
                  },
                  "required": [
                    "detail"
                  ]
                }
              }
            }
          }
        },
        "description": "主检索入口。结果仅为语义候选，不证明颜色、性别等全部条件成立，不自动关联身份。元数据过滤在召回前生效；extra 非空拒绝。故障返回503而非空匹配。"
      }
    },
    "/api/search/person-crops": {
      "post": {
        "tags": [
          "检索"
        ],
        "summary": "严格标签或已知人员检索",
        "operationId": "search_person_crops_api_search_person_crops_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisualSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Basic认证失败；当前响应可为纯文本，并带WWW-Authenticate。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "兼容入口，不做向量兜底。解析出的标签条件要求同时匹配，但不保证自然语言每个词都被解析。rerank 不是启用语义检索开关。"
      }
    },
    "/api/search/index/rebuild": {
      "post": {
        "tags": [
          "检索"
        ],
        "summary": "维护：补建视觉索引",
        "operationId": "rebuild_search_index_api_search_index_rebuild_post",
        "parameters": [
          {
            "name": "target",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(image|person_crop)$",
              "default": "person_crop",
              "title": "Target"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexRebuildResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Basic认证失败；当前响应可为纯文本，并带WWW-Authenticate。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "业务参数无效或媒体无法处理；具体原因见detail。",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "object",
                          "additionalProperties": true
                        },
                        {
                          "type": "array",
                          "items": {}
                        }
                      ]
                    }
                  },
                  "required": [
                    "detail"
                  ]
                }
              }
            }
          }
        },
        "description": "由业务后端或运维受控调用；当前Basic Auth无独立管理员权限区分。按时间倒序处理最近limit条，无offset/单对象参数。必须检查errors与indexed，不只判断200。"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SemanticSearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 20
          },
          "filters": {
            "$ref": "#/components/schemas/SearchFilters"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SemanticSearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SearchResultItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "SearchResponse"
      },
...
      "SearchFilters": {
        "properties": {
          "start_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "camera_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Camera Id"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "extra": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra"
          }
        },
        "type": "object",
        "title": "SearchFilters"
      },
      "PersonCropRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "image_id": {
            "type": "string",
            "format": "uuid",
            "title": "Image Id"
          },
          "crop_url": {
            "type": "string",
            "title": "Crop Url"
          },
          "bbox": {
            "additionalProperties": true,
            "type": "object",
            "title": "Bbox"
          },
          "attributes": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attributes"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "camera_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Camera Id"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "image_id",
          "crop_url",
          "bbox",
          "person_id",
          "camera_id",
          "location_id",
          "captured_at",
          "created_at"
        ],
        "title": "PersonCropRead"
      },
      "SemanticSearchResponse": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "semantic",
            "title": "Mode",
            "default": "semantic"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SemanticSearchItem"
            },
            "type": "array",
            "title": "Items"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "min_score": {
            "type": "number",
            "title": "Min Score"
          },
          "notice": {
            "type": "string",
            "title": "Notice"
          },
          "scope_crops": {
            "type": "integer",
            "title": "Scope Crops"
          },
          "indexed_scope_crops": {
            "type": "integer",
            "title": "Indexed Scope Crops"
          },
          "candidates_examined": {
            "type": "integer",
            "title": "Candidates Examined",
            "default": 0
          },
          "shortlist_limited": {
            "type": "boolean",
            "title": "Shortlist Limited",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "items",
          "model",
          "min_score",
          "notice",
          "scope_crops",
          "indexed_scope_crops"
        ],
        "title": "SemanticSearchResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ImageRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "camera_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Camera Id"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "location_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Name"
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "image_url",
          "thumbnail_url",
          "source_type",
          "camera_id",
          "location_id",
          "captured_at",
          "created_at"
        ],
        "title": "ImageRead"
      },
      "VisualSearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 20
          },
          "target": {
            "type": "string",
            "title": "Target",
            "default": "person_crop"
          },
          "filters": {
            "$ref": "#/components/schemas/SearchFilters"
          },
          "rerank": {
            "type": "boolean",
            "title": "Rerank",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "VisualSearchRequest"
      },
      "IndexRebuildResponse": {
        "properties": {
          "target": {
            "type": "string",
            "title": "Target"
          },
          "requested": {
            "type": "integer",
            "title": "Requested"
          },
          "seen": {
            "type": "integer",
            "title": "Seen"
          },
          "indexed": {
            "type": "integer",
            "title": "Indexed"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "target",
          "requested",
          "seen",
          "indexed"
        ],
        "title": "IndexRebuildResponse"
      },
...
      "SearchResultItem": {
        "properties": {
          "crop_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crop Id"
          },
          "image_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "crop_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crop Url"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "original_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Score"
          },
          "embedding_rerank_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Rerank Score"
          },
          "rerank_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerank Score"
          },
          "rerank_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerank Reason"
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "location_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Name"
          },
          "camera_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Camera Id"
          },
          "camera_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Camera Name"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "person_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Name"
          },
          "attributes": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attributes"
          },
          "labels_zh": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labels Zh"
          },
          "labels_en": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labels En"
          }
        },
        "type": "object",
        "required": [
          "score"
        ],
        "title": "SearchResultItem"
      },
...
      "SemanticSearchStatus": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "configured": {
            "type": "boolean",
            "title": "Configured"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "min_score": {
            "type": "number",
            "title": "Min Score"
          },
          "total_crops": {
            "type": "integer",
            "title": "Total Crops"
          },
          "indexed_crops": {
            "type": "integer",
            "title": "Indexed Crops"
          },
          "labeled_crops": {
            "type": "integer",
            "title": "Labeled Crops"
          },
          "attributes_enabled": {
            "type": "boolean",
            "title": "Attributes Enabled"
          },
          "auto_index_on_ingest": {
            "type": "boolean",
            "title": "Auto Index On Ingest"
          }
        },
        "type": "object",
        "required": [
          "enabled",
          "configured",
          "model",
          "min_score",
          "total_crops",
          "indexed_crops",
          "labeled_crops",
          "attributes_enabled",
          "auto_index_on_ingest"
        ],
        "title": "SemanticSearchStatus"
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "实际部署由中间件执行，账号密码单独配置，本文不包含凭据。"
      }
    }
  },
...
}